requirements.txt
numpy~=1.24.2

python-dotenv~=1.0.0
